src/main/java/com/kumarvv/ketl/utils/Chrono.java
/**
 * Copyright (c) 2016 Vijay Vijayaram
 *
 * Permission is hereby granted, free of charge, to any person obtaining a copy of this software
 * and associated documentation files (the "Software"), to deal in the Software without restriction,
 * including without limitation the rights to use, copy, modify, merge, publish, distribute, sublicense,
 * and/or sell copies of the Software, and to permit persons to whom the Software is furnished to do so,
 * subject to the following conditions:
 *
 * The above copyright notice and this permission notice shall be included in all copies or substantial
 * portions of the Software.
 *
 * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED, INCLUDING BUT
 * NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT.
 * IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY,
 * WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH
 * THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 */
package com.kumarvv.ketl.utils;

import org.pmw.tinylog.Logger;

public class Chrono {
    long millis = 0;
    String name = "";

    public static Chrono start(String name) {
        Chrono ch = new Chrono();
        ch.millis = System.currentTimeMillis();
        ch.name = name;
        return ch;
    }

    public void restart() {
        millis = System.currentTimeMillis();
    }

    public void stop() {
        long diff = System.currentTimeMillis() - millis;
        Logger.info("chrono-[{}] elapsed time: {}ms", name, diff);
    }
}
